+++ NEWS.md
@@ -1,6 +1,7 @@
 # mirai 1.2.0.9001 (development)
 
 * Fixes flatmap with `mirai_map()` assigning the variable 'typ' to the global environment.
+* Progress bars for `mirai_map()` automatically use enhanced versions from the 'cli' package if available.
 
 # mirai 1.2.0